backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
python-docx==1.1.2
python-pptx==1.0.2
anthropic==0.34.2
celery[redis]==5.4.0
redis==5.1.1
cryptography==43.0.1
httpx==0.27.2
aiofiles==24.1.0
